requirements.txt
# Django and core dependencies
Django>=3.2,<5.0
mysqlclient>=2.1,<3.0
python-dotenv>=0.19,<1.0
gunicorn>=20.0,<23.0
python-decouple>=3.4,<4.0

# Django REST framework and JWT
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0,<6.0

# Keycloak integration using mozilla-django-oidc
mozilla-django-oidc>=4.0.0


# Optional: useful for managing CORS (if you're using APIs from frontend)
django-cors-headers>=4.0,<5.0

# Optional: dotenv support for managing environment variables
python-dotenv>=0.19,<1.0
